tee: optee: implement shutdown so kexec can work
The Secure World OP-TEE side caches shm regions it agreed
with the optee kernel driver, and uses them to pass
marshaled arguments for RPC.

In the case that the Normal World OS is going to go away,
eg, for kexec, these cached shm blocks must be invalidated
on the secure OP-TEE OS side.  Otherwise, they will continue
to be used for communication with the new OS.  Since the
shm handle is directly the VA address of the linux driver's
logical shm struct, this results in the new OS driver
dereferencing memory that was only allocated in the old
Linux driver instance.

This patch wires up the .shutdown member of the
platform_driver, which is called before kexec switches
to the new kernel image, and informs OP-TEE to invalidate
any cached shm.

Signed-off-by: Andy Green <andy@warmcat.com>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>